prep for monorepo by moving `pkg/types`, `pkg/testsupport` under `pkg/comments`
These packages will collide with existing packages in
`github.com/weberc2/auth`.
