+++ pkg/comments/commentsservice.go
@@ -6,7 +6,7 @@ import (
 	"net/http"
 	"time"
 
-	"github.com/weberc2/comments/pkg/types"
+	"github.com/weberc2/comments/pkg/comments/types"
 	pz "github.com/weberc2/httpeasy"
 )
